Rename throttle package to limiter
